% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CosMxSpatialAnalysisWorkshop.R
\docType{data}
\name{lung5}
\alias{lung5}
\title{Sub-sampled CosMx non-small cell lung cancer (NSCLC) spatial transcriptomics dataset}
\format{
A SpatialExperiment object.
}
\usage{
lung5
}
\description{
A Bruker CosMx spatial transcriptomics dataset containing a sub-region of a non-small cell lung cancer (NSCLC) sample. The data contains pathology annotations. The full dataset is available through the SubcellularSpatialData R/Bioconductor package.
}
\references{
Bhuva DD, Tan CW, Salim A, Marceaux C, Pickering MA, Chen J, Kharbanda M, Jin X, Liu N, Feher K, Putri G. Library size confounds biology in spatial transcriptomics data. Genome Biology. 2024 Apr 18;25(1):99.

Bhuva DD, Tan CW, Marceaux C, Pickering M, Salim A, Chen J, Kharbanda M, Jin X, Liu N, Feher K, et al. Library size confounds biology in spatial transcriptomics data. 2024. Zenodo. https://doi.org/10.5281/zenodo.7959786.

Bhuva DD: SubcellularSpatialData: annotated spatial transcriptomics datasets from 10x Xenium, NanoString CosMx and BGI STOmics. Bioconductor. 2024 https://doi.org/10.18129/B9.bioc.SubcellularSpatialData.

Bhuva DD. Library size confounds biology in spatial transcriptomics. 2024. Zenodo. https://doi.org/10.5281/zenodo.10946961.
}
\keyword{datasets}
