% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CosMxSpatialAnalysisWorkshop.R
\docType{data}
\name{ref_luca}
\alias{ref_luca}
\title{Pseudo-bulked single-cell non-small cell lung cancer (NSCLC) atlas reference dataset}
\format{
A SingleCellExperiment object.
}
\usage{
ref_luca
}
\description{
A pseudo-bulked single-cell RNA-seq non-small cell lung cancer (NSCLC) atlas dataset from the study by Salcher et al., Cancer Cell, 2022. The original dataset has been subsetted to the genes present in the 10x Xenium dataset.
}
\references{
Salcher S, Sturm G, Horvath L, Untergasser G, Kuempers C, Fotakis G, Panizzolo E, Martowicz A, Trebo M, Pall G, Gamerith G. High-resolution single-cell atlas reveals diversity and plasticity of tissue-resident neutrophils in non-small cell lung cancer. Cancer cell. 2022 Dec 12;40(12):1503-20.
}
\keyword{datasets}
